team: add simple bonding/teaming module
An initial version of a bonding/teaming module based on libteam and the
kernel team driver. It is capable of creating a team device and add
member ports. The ultimate goal is to support LAG offloaded to DSA
switch hardware with teamd handling the LACP management.

    config interface teamdev
        option ifname "lan1 lan2"
        option type team

The module requires libteam's teamd and teamdctl commands.

Signed-off-by: Pavel Šimerda <code@simerda.eu>